server/pydantic-fastapi/pyproject.toml
[tool.poetry]
name = "pydantic-fastapi"
version = "0.1.0"
description = ""
authors = ["HL <myhendry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = {extras = ["standard"], version = "^0.115.8"}
langchain = "^0.3.19"
langchain-core = "^0.3.35"
langchain-community = "^0.3.17"
langchain-qdrant = "^0.2.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-multipart = "^0.0.20"
aiofiles = "^24.1.0"
langchain-openai = "0.3.1"
browser-use = "^0.1.40"
pypdf = "^5.4.0"
cors = "^1.0.1"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
